+++ src/Models/index.ts
@@ -1,0 +1,4 @@
+export { GetClientListResponse } from './getClientListResponse';
+export { ClientList } from './clientList';
+export { SingleClientData } from './singleClientData';
+export { Client } from './client';