+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 -->
 ## **WORK IN PROGRESS**
 * (theimo1221) Refactoring
+* (theomo1221) Implement Basic Client Models
 
 ## 0.0.2-alpha.0 (2022-07-09)
 * (theimo1221) Added Basic Project Structure including Workflow and npm scripts